fuse the prod/dev to enable disable component
