+++ projects/forms/src/lib/show-form-control/show-form-control.component.scss
@@ -1,4 +1,7 @@
 :host {
+  display: none!important;
+}
+:host.enabled {
   display: block;
   position: fixed;
   right: 0;
