+++ projects/forms/src/lib/show-form-control/show-form-control.ts
@@ -1,4 +1,6 @@
-export const componentSelector = 'fty-show-form-control';
+import { InjectionToken } from '@angular/core';
+
+export const componentSelector = 'show-form-control';
 
 export interface WindowWidthProvider {
   innerWidth: number;
@@ -7,3 +9,5 @@ export interface WindowWidthProvider {
 export interface WindowAnimationFrameProvider {
   requestAnimationFrame(cb: any): any;
 }
+
+export const Disable = new InjectionToken<boolean>('Are the form control visualizations disabled? On Prod for example -they should be.');
