+++ src/app/app.component.html
@@ -8,8 +8,8 @@
     [(ngModel)]="person"
     #controlDirective="ngModel"
   ></fty-person-contact>
-  <fty-show-form-control
+  <show-form-control
     [control]="controlDirective.control"
     name="person"
-  ></fty-show-form-control>
+  ></show-form-control>
 </div>
