+++ src/app/party-form/party-form.component.html
@@ -21,4 +21,5 @@
   <button type="reset" color="warn" mat-raised-button>Reset</button>
 </form>
 
-<fty-show-form-control [control]="form" name="Party Form" [closed]="true"></fty-show-form-control>
+<show-form-control [control]="form" name="Party Form" [closed]="true"></show-form-control>
+
